pyproject.toml
[tool.poetry]
name = "cycleman"
version = "0.0.1"
description = "Cycle Management System"
authors = ["Abhinav Kumar <abhinav.vikash2@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
django-debug-toolbar = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
